move more sound structs from arm9 to common include
